Extract config out
