+++ packages/bundlers/default/src/DefaultBundler.js
@@ -6,75 +6,22 @@ import type {
   BundleGroup,
   Dependency,
   Environment,
-  Config,
   MutableBundleGraph,
-  PluginOptions,
   Target,
-  BuildMode,
   PluginLogger,
 } from '@atlaspack/types';
 import type {NodeId} from '@atlaspack/graph';
-import type {SchemaEntity} from '@atlaspack/utils';
 import {ContentGraph, Graph, BitSet, ALL_EDGE_TYPES} from '@atlaspack/graph';
 
 import invariant from 'assert';
 import {Bundler} from '@atlaspack/plugin';
-import {validateSchema, DefaultMap, globToRegex} from '@atlaspack/utils';
+import {DefaultMap, globToRegex} from '@atlaspack/utils';
 import nullthrows from 'nullthrows';
 import path from 'path';
-import {encodeJSONKeyComponent} from '@atlaspack/diagnostic';
 import {getFeatureFlag} from '@atlaspack/feature-flags';
+
+import {loadBundlerConfig, type ResolvedBundlerConfig} from './bundlerConfig';
 import {addJSMonolithBundle} from './MonolithicBundler';
-
-type Glob = string;
-
-type ManualSharedBundles = Array<{|
-  name: string,
-  assets: Array<Glob>,
-  types?: Array<string>,
-  root?: string,
-  split?: number,
-|}>;
-
-type BaseBundlerConfig = {|
-  http?: number,
-  minBundles?: number,
-  minBundleSize?: number,
-  maxParallelRequests?: number,
-  disableSharedBundles?: boolean,
-  manualSharedBundles?: ManualSharedBundles,
-|};
-
-type BundlerConfig = {|
-  [mode: BuildMode]: BaseBundlerConfig,
-|} & BaseBundlerConfig;
-
-type ResolvedBundlerConfig = {|
-  minBundles: number,
-  minBundleSize: number,
-  maxParallelRequests: number,
-  projectRoot: string,
-  disableSharedBundles: boolean,
-  manualSharedBundles: ManualSharedBundles,
-|};
-
-// Default options by http version.
-const HTTP_OPTIONS = {
-  '1': {
-    minBundles: 1,
-    manualSharedBundles: [],
-    minBundleSize: 30000,
-    maxParallelRequests: 6,
-    disableSharedBundles: false,
-  },
-  '2': {
-    minBundles: 1,
-    manualSharedBundles: [],
-    minBundleSize: 20000,
-    maxParallelRequests: 25,
-    disableSharedBundles: false,
-  },
-};
 
 /* BundleRoot - An asset that is the main entry of a Bundle. */
 type BundleRoot = Asset;
@@ -1629,60 +1576,6 @@ function createIdealGraph(
   };
 }
 
-const CONFIG_SCHEMA: SchemaEntity = {
-  type: 'object',
-  properties: {
-    http: {
-      type: 'number',
-      enum: Object.keys(HTTP_OPTIONS).map((k) => Number(k)),
-    },
-    manualSharedBundles: {
-      type: 'array',
-      items: {
-        type: 'object',
-        properties: {
-          name: {
-            type: 'string',
-          },
-          assets: {
-            type: 'array',
-            items: {
-              type: 'string',
-            },
-          },
-          types: {
-            type: 'array',
-            items: {
-              type: 'string',
-            },
-          },
-          root: {
-            type: 'string',
-          },
-          split: {
-            type: 'number',
-          },
-        },
-        required: ['name', 'assets'],
-        additionalProperties: false,
-      },
-    },
-    minBundles: {
-      type: 'number',
-    },
-    minBundleSize: {
-      type: 'number',
-    },
-    maxParallelRequests: {
-      type: 'number',
-    },
-    disableSharedBundles: {
-      type: 'boolean',
-    },
-  },
-  additionalProperties: false,
-};
-
 function createBundle(opts: {|
   uniqueKey?: string,
   target: Target,
@@ -1722,121 +1615,6 @@ function createBundle(opts: {|
     needsStableName: Boolean(opts.needsStableName),
     bundleBehavior: opts.bundleBehavior ?? asset.bundleBehavior,
     manualSharedBundle: opts.manualSharedBundle,
-  };
-}
-
-function resolveModeConfig(
-  config: BundlerConfig,
-  mode: BuildMode,
-): BaseBundlerConfig {
-  let generalConfig = {};
-  let modeConfig = {};
-
-  for (const key of Object.keys(config)) {
-    if (key === 'development' || key === 'production') {
-      if (key === mode) {
-        modeConfig = config[key];
-      }
-    } else {
-      generalConfig[key] = config[key];
-    }
-  }
-
-  // $FlowFixMe Not sure how to convince flow here...
-  return {
-    ...generalConfig,
-    ...modeConfig,
-  };
-}
-
-async function loadBundlerConfig(
-  config: Config,
-  options: PluginOptions,
-  logger: PluginLogger,
-): Promise<ResolvedBundlerConfig> {
-  let conf = await config.getConfig<BundlerConfig>([], {
-    packageKey: '@atlaspack/bundler-default',
-  });
-
-  if (!conf) {
-    const modDefault = {
-      ...HTTP_OPTIONS['2'],
-      projectRoot: options.projectRoot,
-    };
-    return modDefault;
-  }
-
-  invariant(conf?.contents != null);
-
-  let modeConfig = resolveModeConfig(conf.contents, options.mode);
-
-  // minBundles will be ignored if shared bundles are disabled
-  if (
-    modeConfig.minBundles != null &&
-    modeConfig.disableSharedBundles === true
-  ) {
-    logger.warn({
-      origin: '@atlaspack/bundler-default',
-      message: `The value of "${modeConfig.minBundles}" set for minBundles will not be used as shared bundles have been disabled`,
-    });
-  }
-
-  // minBundleSize will be ignored if shared bundles are disabled
-  if (
-    modeConfig.minBundleSize != null &&
-    modeConfig.disableSharedBundles === true
-  ) {
-    logger.warn({
-      origin: '@atlaspack/bundler-default',
-      message: `The value of "${modeConfig.minBundleSize}" set for minBundleSize will not be used as shared bundles have been disabled`,
-    });
-  }
-
-  // maxParallelRequests will be ignored if shared bundles are disabled
-  if (
-    modeConfig.maxParallelRequests != null &&
-    modeConfig.disableSharedBundles === true
-  ) {
-    logger.warn({
-      origin: '@atlaspack/bundler-default',
-      message: `The value of "${modeConfig.maxParallelRequests}" set for maxParallelRequests will not be used as shared bundles have been disabled`,
-    });
-  }
-
-  if (modeConfig.manualSharedBundles) {
-    let nameArray = modeConfig.manualSharedBundles.map((a) => a.name);
-    let nameSet = new Set(nameArray);
-    invariant(
-      nameSet.size == nameArray.length,
-      'The name field must be unique for property manualSharedBundles',
-    );
-  }
-
-  validateSchema.diagnostic(
-    CONFIG_SCHEMA,
-    {
-      data: modeConfig,
-      source: await options.inputFS.readFile(conf.filePath, 'utf8'),
-      filePath: conf.filePath,
-      prependKey: `/${encodeJSONKeyComponent('@atlaspack/bundler-default')}`,
-    },
-    '@atlaspack/bundler-default',
-    'Invalid config for @atlaspack/bundler-default',
-  );
-
-  let http = modeConfig.http ?? 2;
-  let defaults = HTTP_OPTIONS[http];
-
-  return {
-    minBundles: modeConfig.minBundles ?? defaults.minBundles,
-    minBundleSize: modeConfig.minBundleSize ?? defaults.minBundleSize,
-    maxParallelRequests:
-      modeConfig.maxParallelRequests ?? defaults.maxParallelRequests,
-    projectRoot: options.projectRoot,
-    disableSharedBundles:
-      modeConfig.disableSharedBundles ?? defaults.disableSharedBundles,
-    manualSharedBundles:
-      modeConfig.manualSharedBundles ?? defaults.manualSharedBundles,
   };
 }
